interfaz funcional

+++ common/include/update.h
@@ -9,18 +9,18 @@ class Update{
 private:
 	uint32_t idObject;
 	STATUS status;
-	float posX;
-	float posY;
+	int32_t posX;
+	int32_t posY;
 	uint32_t direction;
 public:
 	Update();
-	explicit Update(uint32_t id, STATUS status, float posx,
-	float posy, uint32_t dir);
+	explicit Update(uint32_t id, STATUS status, int32_t posx,
+	int32_t posy, uint32_t dir);
 	~Update();
 	uint32_t getId() const;
 	STATUS getStatus() const;
-	float getPosX() const;
-	float getPosY() const;
+	int32_t getPosX() const;
+	int32_t getPosY() const;
 	uint32_t getDirection() const;
 };
