+++ common/src/update.cpp
@@ -2,8 +2,8 @@
 
 Update::Update() : idObject(0),status(NONE), posX(0), posY(0), direction(0){}
 
-Update::Update(uint32_t id, STATUS status, float posx,
-	float posy, uint32_t dir) 
+Update::Update(uint32_t id, STATUS status, int32_t posx,
+	int32_t posy, uint32_t dir) 
 : idObject(id),status(status), posX(posx), posY(posy), direction(dir) {}
 
 Update::~Update(){}
@@ -16,11 +16,11 @@ STATUS Update::getStatus() const{
 	return this->status;
 }
 
-float Update::getPosX() const{
+int32_t Update::getPosX() const{
 	return this->posX;
 }
 
-float Update::getPosY() const{
+int32_t Update::getPosY() const{
 	return this->posY;
 }
